create auth

+++ controllers/cards.js
@@ -1,51 +1,51 @@
+const PageNotFoundError = require('../errors/PageNotFoundError');
+const UserNotFoundError = require('../errors/UserNotFoundError');
+const ValidationError = require('../errors/ValidationError');
 const CardModel = require('../models/card');
 
 const STATUS_OK = 200;
-const ERROR_VALIDATION = 400;
-const ERROR_NOT_FOUND = 404;
-const ERROR_SERVER = 500;
+const STATUS_CREATED = 201;
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   CardModel.create({ name, link, owner: req.user._id })
-    .then((data) => res.status(201).send(data))
+    .then((data) => res.status(STATUS_CREATED).send(data))
     .catch((err) => {
-      console.log(err);
       if (err.name === 'ValidationError') {
-        return res.status(ERROR_VALIDATION).send({ message: `Переданы некорректные данные, ${err}` });
+        return next(new ValidationError('Переданы некорректные данные'));
       }
-      return res.status(ERROR_SERVER).send({ message: `Ошибка на стороне сервера, ${err}` });
+      return next(err);
     });
 };
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   CardModel.find()
     .then((cards) => res.status(STATUS_OK).send(cards))
-    .catch((err) => {
-      console.log(err);
-      return res.status(ERROR_SERVER).send({ message: `Ошибка на стороне сервера, ${err}` });
-    });
+    .catch((err) => next(err));
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  CardModel.findByIdAndRemove(cardId)
-    .then((card) => {
-      if (!card) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка отсутствует' });
+  CardModel.findById(cardId)
+    .then((deletedCard) => {
+      if (!deletedCard) {
+        return next(new PageNotFoundError('Запрашиваемая карточка отсутствует'));
       }
-      return res.status(STATUS_OK).send(card);
+      if (deletedCard.owner.toString() !== req.user._id) {
+        return next(new UserNotFoundError('Нет прав для удаления карточки другого пользователя'));
+      }
+      res.status(STATUS_OK).send({ message: 'Карточка удалена' });
+      return CardModel.findByIdAndRemove(cardId);
     })
     .catch((err) => {
-      console.log(err);
       if (err.name === 'CastError') {
-        return res.status(ERROR_VALIDATION).send({ message: `Переданы некорректные данные, ${err}` });
+        return next(new ValidationError('Переданы некорректные данные'));
       }
-      return res.status(ERROR_SERVER).send({ message: `Ошибка на стороне сервера, ${err}` });
+      return next(err);
     });
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   CardModel.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -53,20 +53,19 @@ const likeCard = (req, res) => {
   )
     .then((data) => {
       if (!data) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка отсутствует' });
+        return next(new PageNotFoundError('Запрашиваемая карточка отсутствует'));
       }
       return res.status(STATUS_OK).send(data);
     })
     .catch((err) => {
-      console.log(err);
       if (err.name === 'CastError') {
-        return res.status(ERROR_VALIDATION).send({ message: `Переданы некорректные данные, ${err}` });
+        return next(new ValidationError('Переданы некорректные данные'));
       }
-      return res.status(ERROR_SERVER).send({ message: `Ошибка на стороне сервера, ${err}` });
+      return next(err);
     });
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   CardModel.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -74,16 +73,15 @@ const dislikeCard = (req, res) => {
   )
     .then((data) => {
       if (!data) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка отсутствует' });
+        return next(new PageNotFoundError('Запрашиваемая карточка отсутствует'));
       }
       return res.status(STATUS_OK).send(data);
     })
     .catch((err) => {
-      console.log(err);
       if (err.name === 'CastError') {
-        return res.status(ERROR_VALIDATION).send({ message: `Переданы некорректные данные, ${err}` });
+        return next(new ValidationError('Переданы некорректные данные'));
       }
-      return res.status(ERROR_SERVER).send({ message: `Ошибка на стороне сервера, ${err}` });
+      return next(err);
     });
 };
